Summer Worth of Changes
I have added a bunch of new features/fixes to my code along with more files that the program uses/used for testing purposes.

such as:
enum Enum (used for manual or Automatic radio buttons)

struct OpenReportInfo (used for saving the currently opened report and its data to update the list)

struct AppConfigs (used for getting all of the AppConfigs.json data and parsing it to be used in the program)

struct NewReportValues (used when creating a new report for what tests to run and if they are automatic or manual)

impl AppConfigs (used to parse the .json appConfigs and updating the data within that file. Namely the recent files for right now.)

All other details of changes can be seen in when looking at the changes made. To much to type out lol XD